server/core/requirements.txt
django==4.2.7
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
requests==2.31.0
python-dotenv==1.0.0
